broke out Entry and IconSet into UI components for import
